Move the stream salvage into its own mixin, alongside the param headers
The branch below added the capture middleware and the salvage that reads
its buffer, which is what stops a break after the final event from
re-running a completed call. It is a unit, and moving it keeps the base
module under its length cop the same way the parameter headers already
did.
